Update Inputs.md
Sorry it must have been an autocomplete mistake for the 'addition'. See new commit.

Out of curiosity I search for 'edition' and if it is can be actually used in our case but I could not find something. So it might be better to express it in this way as shown in the new commit.